requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
Werkzeug==3.0.6
reportlab==4.2.5
weasyprint==62.3
pandas==2.2.3
numpy==1.24.4
matplotlib==3.7.3
seaborn==0.13.2
plotly==5.24.1
requests==2.32.3
Pillow==11.1.0